Phishing Simulation Report/requirements.txt
Markdown==3.8.2
matplotlib==3.10.3
numpy==2.3.1
pandas==2.3.1
protobuf==6.31.1
python-dotenv==1.1.1
reportlab==4.4.2
google-generativeai